chore: Move the Cta link to it's own component
