+++ package.json
@@ -24,6 +24,7 @@
     "@contentful/rich-text-types": "^14.1.0",
     "@reach/alert": "^0.9.1",
     "@reach/combobox": "^0.9.1",
+    "@reach/disclosure": "^0.10.2",
     "@reach/menu-button": "^0.10.2",
     "@reach/router": "^1.3.3",
     "@reach/skip-nav": "^0.9.0",
