+++ src/components/common/landing-page/call-to-action.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Link } from 'gatsby'
-import ctaLinkStyle from './cta-link.module.scss'
+import ctaLinkStyle from './call-to-action.module.scss'
 
-export default ({ to, children, centered = false, bold = false }) => (
+const CtaLink = ({ to, children, centered = false, bold = false }) => (
   <Link
     to={to}
     className={`${ctaLinkStyle.cta} ${centered ? ctaLinkStyle.centered : ''} ${
@@ -16,3 +16,7 @@ export default ({ to, children, centered = false, bold = false }) => (
     </span>
   </Link>
 )
+
+export default CtaLink
+
+export { CtaLink }
