+++ src/components/pages/homepage/blog-list.js
@@ -3,7 +3,7 @@ import { Link, useStaticQuery, graphql } from 'gatsby'
 import pressListStyle from '~components/common/press-list.module.scss'
 import { Byline } from '~components/pages/blog/byline'
 import Container from '~components/common/landing-page/container'
-import CtaLink from '~components/common/landing-page/cta-link'
+import { CtaLink } from '~components/common/landing-page/call-to-action'
 import Paragraph from '~components/common/landing-page/paragraph'
 import blogListStyles from './blog-list.module.scss'
