+++ src/components/pages/race/charts/counties-table.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { Table, Th, Td } from '~components/common/table'
 import countiesTableStyle from './counties-table.module.scss'
-import CtaLink from '~components/common/landing-page/cta-link'
+import { CtaLink } from '~components/common/landing-page/call-to-action'
 import { FormatNumber } from '~components/utils/format'
 
 const CountyTable = ({ tableSource, defaultSort, getRank }) => {
